third_party/blink/web_tests/platform/win/external/wpt/html/semantics/forms/the-select-element/customizable-select/switch-picker-appearance-expected.txt

This is a testharness.js-based test.
[FAIL] Basic functionality of select picker with appearance:auto
  assert_equals: appearance:auto picker is never open expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] Basic functionality of select picker with appearance:none
  assert_equals: appearance:none picker is never open expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
[FAIL] Switching appearance in :open should close the picker
  assert_equals: expected "base-select" but got "auto"
Harness: the test ran to completion.
